modified:   src/Edge.cpp 	modified:   src/Edge.h 	modified:   src/Graph.cpp 	modified:   src/Graph.h 	modified:   src/Path.cpp 	modified:   src/Path.h

+++ src/Edge.cpp
@@ -1,6 +1,6 @@
 #include "Edge.h"
 
-Edge::Edge(short from, short to, short weight, short ID)
+Edge::Edge(unsigned int from, unsigned int to, short weight, short ID)
 {
     this->fromID = fromID;
     this->toID = toID;
