+++ src/Path.cpp
@@ -1,6 +1,6 @@
 #include "Path.h"
 
-Path::Path(short sourceNodeID, std::vector<Edge> &edgesToDestination)
+Path::Path(unsigned int sourceNodeID, std::vector<Edge> &edgesToDestination)
 {
     this->sourceNodeID = sourceNodeID;
     this->edgesToDestination = edgesToDestination;
